+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Release 0.1.2
+
+- [x] Swift 4.2
+- [x] SFSafariViewController compatible.
+- [x] UIViewController extension.
+
 # Release 0.1.1
 
 - [x] Swift 4.1